# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 2.2.0
  description: Construction AI Platform — Phase 2 API (routes match FastAPI backend exactly)
servers:
  - url: /api
    description: Base API path (Python FastAPI backend)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Notification:
      type: object
      required: [id, organization_id, recipient_user_id, event_type, entity_type, entity_id,
                 title, message, severity, is_read, created_at]
      properties:
        id:
          type: integer
        organization_id:
          type: integer
        recipient_user_id:
          type: integer
        actor_user_id:
          type: ["integer", "null"]
        project_id:
          type: ["integer", "null"]
        event_type:
          type: string
        entity_type:
          type: string
        entity_id:
          type: integer
        title:
          type: string
        message:
          type: string
        severity:
          type: string
        action_url:
          type: ["string", "null"]
        is_read:
          type: boolean
        read_at:
          type: ["string", "null"]
          format: date-time
        created_at:
          type: string
          format: date-time
...
    ProjectMembership:
      type: object
      required: [id, user_id, project_id, role_on_project, is_active, created_at]
      properties:
        id:
          type: integer
        user_id:
          type: integer
        project_id:
          type: integer
        role_on_project:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
security:
  - bearerAuth: []
paths:
...
  /v1/projects/{project_id}/memberships:
    get:
      operationId: listProjectMemberships
      tags: [projects]
      summary: List project memberships (used to resolve assignable users)
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of project memberships
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectMembership"
...
  /v1/notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List the caller's own notifications
      parameters:
        - name: unread_only
          in: query
          schema:
            type: boolean
            default: false
        - name: event_type
          in: query
          schema:
            type: string
        - name: project_id
          in: query
          schema:
            type: integer
        - name: severity
          in: query
          schema:
            type: string
        - name: created_after
          in: query
          schema:
            type: string
            format: date-time
        - name: created_before
          in: query
          schema:
            type: string
            format: date-time
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
...
  /v1/notifications/{notification_id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a single notification read
      parameters:
        - name: notification_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "404":
          description: Not found
...